Fix round robin scheduler

+++ t1/escalonador_rr.c
@@ -2,6 +2,8 @@
 #include <signal.h>
 #include <unistd.h>
 #include <sys/stat.h>
+#include <sys/types.h>
+#include <sys/wait.h>
 #include <string.h>
 #include "interpreter.h"
 #define INTERVAL 1 //ms
@@ -13,10 +15,8 @@ void remove_from_array(Process *p_pool, int index, int size) {
 }   
 
 int pid_exists(int pid) {
-    struct stat sts;
-    char pid_path[16];
-    sprintf(pid_path, "/proc/%d", pid);
-    return stat(pid_path, &sts) == 0;
+    int status;
+    return pid != waitpid(pid, &status, WNOHANG);
 }
 
 int main(void) {
@@ -33,10 +33,18 @@ int main(void) {
         
         prev = cur > 0 ? cur - 1 : num_process - 1;
       
-        kill(p_pool[prev].pid, SIGSTOP);
+        if (num_process > 1) {
+            kill(p_pool[prev].pid, SIGSTOP);
+        }        
 
-        kill(p_pool[cur].pid, SIGCONT);
-
+        if (pid_exists(p_pool[cur].pid)) {
+            kill(p_pool[cur].pid, SIGCONT);
+        } else {
+            remove_from_array(p_pool, cur, num_process);
+            num_process--;
+            cur--;
+        }
+        
 
         sleep(INTERVAL);
         cur++;
